requirements.txt
streamlit
langchain
langchain-community
openai
python-docx
pypdf
sentence-transformers
chromadb
python-dotenv
tiktoken
